# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API de consultas ao DETRAN-ES",
        "version": "0.0.1",
        "title": "api-detran"
    },
    "basePath": "/",
    "tags": [
        {
            "name": "api-detran",
            "description": ""
        }
    ],
    "schemes": [
        "https"
    ],
    "paths": {
        "/veiculos/{placa}/{doc_proprietario}": {
            "get": {
                "summary": "Dados do veiculo",
                "description": "retorna os dados do veiculo através do WebService InternetBanking",
                "parameters": [
                    {
                        "name": "doc_proprietario",
                        "required": true,
                        "in": "path",
                        "description": "Documento do proprietario do veiculo: CPF ou CNPJ",
                        "type": ""
                    },
                    {
                        "name": "placa",
                        "required": true,
                        "in": "path",
                        "description": "Placa do veiculo",
                        "type": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna informações do veiculo ",
                        "schema": {
                            "$ref": "#/definitions/VeiculoRetorno"
                        }
                    },
                    "403": {
                        "description": "Retorna uma MensagemErro"
                    }
                },
                "tags": [
                    "api-detran"
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ]
            }
        },
        "/veiculos/{placa}/{doc_proprietario}/debitos": {
            "get": {
                "summary": "Débitos do veiculo",
                "description": "retorna uma lista com os débitos do veiculo",
                "parameters": [
                    {
                        "name": "doc_proprietario",
                        "required": true,
                        "in": "path",
                        "description": "Documento do proprietario do veiculo: CPF ou CNPJ",
                        "type": ""
                    },
                    {
                        "name": "placa",
                        "required": true,
                        "in": "path",
                        "description": "Placa do veiculo",
                        "type": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Veiculo encontrado, retorna um array de debitos",
                        "schema": {
                            "$ref": "#/definitions/Debito"
                        }
                    },
                    "403": {
                        "description": "Retorna uma MensagemErro"
                    }
                },
                "tags": [
                    "api-detran"
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ]
            }
        },
        "/veiculos/{placa}/{doc_proprietario}/debitos-preview": {
            "get": {
                "summary": "Prévia dos débitos do veiculo",
                "description": "Retorna uma previa dos débitos do veiculo",
                "parameters": [
                    {
                        "name": "doc_proprietario",
                        "required": true,
                        "in": "path",
                        "description": "Documento do proprietario do veiculo: CPF ou CNPJ",
                        "type": ""
                    },
                    {
                        "name": "placa",
                        "required": true,
                        "in": "path",
                        "description": "Placa do veiculo",
                        "type": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Veiculo encontrado, retorna um array de debitos",
                        "schema": {
                            "$ref": "#/definitions/Debito"
                        }
                    },
                    "403": {
                        "description": "Retorna uma MensagemErro"
                    }
                },
                "tags": [
                    "api-detran"
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ]
            }
        },
        "/veiculos/{placa}/{doc_proprietario}/debitos-tipo/{tipo_debito}": {
            "get": {
                "summary": "Prévia dos débitos do veiculo",
                "description": "Retorna uma lista de um tipo de débitos do veiculo",
                "parameters": [
                    {
                        "name": "tipo_debito",
                        "required": true,
                        "in": "path",
                        "description": "Tipo de debitos",
                        "type": ""
                    },
                    {
                        "name": "doc_proprietario",
                        "required": true,
                        "in": "path",
                        "description": "Documento do proprietario do veiculo: CPF ou CNPJ",
                        "type": ""
                    },
                    {
                        "name": "placa",
                        "required": true,
                        "in": "path",
                        "description": "Placa do veiculo",
                        "type": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Veiculo encontrado, retorna um array de debitos",
                        "schema": {
                            "$ref": "#/definitions/Debito"
                        }
                    },
                    "403": {
                        "description": "Retorna uma MensagemErro"
                    }
                },
                "tags": [
                    "api-detran"
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ]
            }
        },
        "/veiculos/{placa}/{doc_proprietario}/gerar-gru": {
            "get": {
                "summary": "Gerar GRU de todosos débitos",
                "description": "Retornar uma GRU com todos os debitos ",
                "parameters": [
                    {
                        "name": "doc_proprietario",
                        "required": true,
                        "in": "path",
                        "description": "Documento do proprietario do veiculo: CPF ou CNPJ",
                        "type": ""
                    },
                    {
                        "name": "placa",
                        "required": true,
                        "in": "path",
                        "description": "Placa do veiculo",
                        "type": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Veiculo encontrado, retorna o um array de itens com o pdf do boleto"
                    },
                    "403": {
                        "description": "Retorna uma MensagemErro"
                    }
                },
                "tags": [
                    "api-detran"
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ]
            }
        }
    },
    "definitions": {
        "VeiculoRetorno": {
            "type": "object",
            "properties": {
                "placa": {
                    "type": "string"
                },
                "modelo": {
                    "type": "string"
                },
                "renavam": {
                    "type": "number"
                }
            },
            "required": [
                "placa",
                "modelo",
                "renavam"
            ]
        },
        "Debito": {
            "type": "object",
            "properties": {
                "classe": {
                    "type": "number"
                },
                "codigoServico": {
                    "type": "number"
                },
                "descricaoServico": {
                    "type": "string"
                },
                "dataVencimento": {
                    "type": "string"
                },
                "exercicio": {
                    "type": "number"
                },
                "idDebito": {
                    "type": "number"
                },
                "parcela": {
                    "type": "number"
                },
                "placa": {
                    "type": "string"
                },
                "valorAtualizadoFranquia": {
                    "type": "number"
                },
                "dpvatAnterior": {
                    "type": "number"
                },
                "dpvatCotas": {
                    "type": "string"
                },
                "dpvatExercicio": {
                    "type": "number"
                },
                "ipvaAnterior": {
                    "type": "number"
                },
                "ipvaExercicio": {
                    "type": "number"
                },
                "ipvaParcelamento": {
                    "type": "number"
                },
                "licenciamentoAnterior": {
                    "type": "number"
                },
                "licenciamentoExercicio": {
                    "type": "number"
                },
                "multas": {
                    "type": "number"
                },
                "taxaEspecial": {
                    "type": "number"
                },
                "taxaPatio": {
                    "type": "number"
                },
                "taxaServico": {
                    "type": "number"
                },
                "ipvaCotas": {
                    "type": "string"
                }
            },
            "required": [
                "classe",
                "codigoServico",
                "descricaoServico",
                "dataVencimento",
                "exercicio",
                "idDebito",
                "parcela",
                "placa",
                "valorAtualizadoFranquia",
                "dpvatAnterior",
                "dpvatCotas",
                "dpvatExercicio",
                "ipvaAnterior",
                "ipvaExercicio",
                "ipvaParcelamento",
                "licenciamentoAnterior",
                "licenciamentoExercicio",
                "multas",
                "taxaEspecial",
                "taxaPatio",
                "taxaServico",
                "ipvaCotas"
            ]
        }
    }
}